[fix] fix usability for low bandwidth
Separated the entry loading and the entry refreshing. Concurrency behavior is now set as: if user is not editing (ie. not keyword inputs during 20 secondes) the refresh will pull the database and refresh the entry. If user is editing, the refresh will force push the database erasing changes. In thi case, a journal opened somewhere will stay-up to date and not erasing database when modified later.